requirements.txt
# OfficeGuard - 系统优化助手
# Python 依赖包

# 系统托盘支持
pystray>=0.19.0

# 图像处理（托盘图标生成）
Pillow>=10.0.0

# 全局快捷键监听
pynput>=1.7.0

# 打包工具（开发环境需要）
pyinstaller>=6.0.0
